hapi/requirements.txt
flask
azure-cosmos
python-dotenv
